Added comments to Price Module.

+++ price_module/contracts/price/ChainlinkService.sol
@@ -2,15 +2,20 @@
 pragma solidity >=0.5.0 <0.7.0;
 import "@chainlink/contracts/src/v0.5/interfaces/AggregatorV3Interface.sol";
 
-contract ChainlinkService {  
-  
-    function getLatestPrice(address feedAddress) 
-        public 
-        view 
-        returns (int, uint, uint8) 
+contract ChainlinkService {
+    /// @dev Function to get price information of a token from using chainlink price feeds.
+    /// @param feedAddress Feed address of the token.
+    function getLatestPrice(address feedAddress)
+        public
+        view
+        returns (
+            int256,
+            uint256,
+            uint8
+        )
     {
         AggregatorV3Interface priceFeed = AggregatorV3Interface(feedAddress);
-        ( ,int price, ,uint timeStamp, ) = priceFeed.latestRoundData();
+        (, int256 price, , uint256 timeStamp, ) = priceFeed.latestRoundData();
         uint8 decimal = priceFeed.decimals();
         return (price, timeStamp, decimal);
     }
